Remove DIY cache system
This branch removes osmdroid's caching mechanisms in favor of using
OkHttp's cache. This means we don't have to maintain a cache, and it
also means that we can cache things transparently. The osmdroid cache
also was 2 step:

http request -> file
file -> map

I don't think this could be very good for performance. Right now
that 2-step is not removed yet, so no tiles appear on this branch.
That is the task to do.